+++ packages/integration-sdk-cli/src/commands/collect.ts
@@ -21,7 +21,7 @@ export function collect() {
     .description('collect data and store entities and relationships to disk')
     .option(
       '-p, --project-path <directory>',
-      'absolute path to integration project directory',
+      'path to integration project directory',
       process.cwd(),
     )
     .option(